Normalize PATCHGET file extensions to lowercase
