+++ Core/Tools/PATCHGET/CHATAPI.h
@@ -18,17 +18,30 @@
 
 #pragma once
 
-#define WIN32_LEAN_AND_MEAN
+#include "COMINIT.h"
+#include <stdio.h>
 #include <windows.h>
-#include <windowsx.h>
-#include <stdlib.h>
+#include <commctrl.h>
+#include <winerror.h>
+#include <ocidl.h>
+#include <olectl.h>
 
-#include"WSTYPES.H"
+/**********************************************************************
+**	This macro serves as a general way to determine the number of elements
+**	within an array.
+*/
+#define	ARRAY_SIZE(x)		int(sizeof(x)/sizeof(x[0]))
+#define size_of(typ,id) sizeof(((typ*)0)->id)
 
-  extern HINSTANCE Global_instance;
-  extern LPSTR     Global_commandline;
-  extern int       Global_commandshow;
+namespace patchget
+{
 
-  extern int       main(int argc, char *argv[]);
+int main(int argc, char *argv[]);
 
-  int              Print_WM(UINT wm,char *out);
+void Startup_Chat(void);
+void Shutdown_Chat(void);
+void Update_If_Required(void);
+
+char const * Fetch_String(int id);
+
+} // namespace patchget
