+++ Core/Tools/PATCHGET/DownloadManager.cpp
@@ -21,9 +21,9 @@
 // Author: Matthew D. Campbell, July 2002
 
 #include "debug.h"
-#include "CHATAPI.H"
+#include "CHATAPI.h"
 #include "DownloadManager.h"
-#include "RESOURCE.H"
+#include "RESOURCE.h"
 
 namespace patchget
 {
