+++ Core/Tools/PATCHGET/RESOURCE.h
