+++ Core/Tools/PATCHGET/WSTYPES.h
